requirements.txt
streamlit
polars
httpx
plotly
python-dotenv
